+++ linkerd/telemeter/usage/src/main/scala/io/buoyant/linkerd/telemeter/UsageDataTelemeter.scala
@@ -8,7 +8,6 @@ import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.scala.DefaultScalaModule
 import com.fasterxml.jackson.module.scala.experimental.ScalaObjectMapper
 import com.google.protobuf.CodedOutputStream
-import com.twitter.common.metrics.Metrics
 import com.twitter.conversions.time._
 import com.twitter.finagle._
 import com.twitter.finagle.http.{MediaType, Method, Request, Response}
@@ -24,7 +23,7 @@ import io.buoyant.linkerd.Linker.param.LinkerConfig
 import io.buoyant.linkerd.protocol.HttpConfig
 import io.buoyant.linkerd.usage.{Counter, Gauge, Router, UsageMessage}
 import io.buoyant.linkerd.{Build, Linker}
-import io.buoyant.telemetry.{Telemeter, TelemeterConfig, TelemeterInitializer}
+import io.buoyant.telemetry._
 import java.nio.ByteBuffer
 import java.text.SimpleDateFormat
 import java.util.{Date, TimeZone}
@@ -41,7 +40,7 @@ private[telemeter] object UsageDataTelemeter {
     pid: String,
     orgId: Option[String]
   ) {
-    def apply(metrics: Map[String, Number]): Future[Unit] = {
+    def apply(metrics: MetricsTree): Future[Unit] = {
       val msg = mkUsageMessage(config, pid, orgId, metrics)
       log.debug(msg.toString)
 
@@ -62,7 +61,7 @@ private[telemeter] object UsageDataTelemeter {
     config: Linker.LinkerConfig,
     pid: String,
     orgId: Option[String],
-    metrics: Map[String, Number]
+    metrics: MetricsTree
   ): UsageMessage =
     UsageMessage(
       pid = Some(pid),
@@ -81,29 +80,42 @@ private[telemeter] object UsageDataTelemeter {
   def mkContainerManager: Option[String] =
     sys.env.get("MESOS_TASK_ID").map(_ => "mesos")
 
-  def mkStartTime(metrics: Map[String, Number]): Option[String] = {
+  def mkStartTime(metrics: MetricsTree): Option[String] = {
     val tz = TimeZone.getTimeZone("UTC")
     val df = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm'Z'")
     df.setTimeZone(tz)
-    val uptime = metrics.get("jvm/uptime").map(_.intValue()).getOrElse(0)
+    val uptime = gaugeValue(metrics.resolve(Seq("jvm", "uptime")).metric).map(_.toInt).getOrElse(0)
     Some(df.format(new Date(System.currentTimeMillis() - uptime)))
   }
 
   def mkNamers(config: Linker.LinkerConfig): Seq[String] =
     config.namers.getOrElse(Seq()).map(_.kind)
 
-  def mkGauges(metrics: Map[String, Number]): Seq[Gauge] =
+  def gaugeValue(metric: Metric): Option[Double] = metric match {
+    case g: Metric.Gauge => Some(g.get)
+    case _ => None
+  }
+
+  def mkGauges(metrics: MetricsTree): Seq[Gauge] =
     Seq(
-      Gauge(Some("jvm_mem"), metrics.get("jvm/mem/current/used").map(_.doubleValue())),
-      Gauge(Some("jvm/gc/msec"), metrics.get("jvm/mem/current/used").map(_.doubleValue())),
-      Gauge(Some("jvm/uptime"), metrics.get("jvm/uptime").map(_.doubleValue()))
+      Gauge(Some("jvm_mem"), gaugeValue(metrics.resolve(Seq("jvm", "mem", "current", "used")).metric)),
+      Gauge(Some("jvm/gc/msec"), gaugeValue(metrics.resolve(Seq("jvm", "mem", "current", "used")).metric)),
+      Gauge(Some("jvm/uptime"), gaugeValue(metrics.resolve(Seq("jvm", "uptime")).metric))
     )
 
-  val RequestPattern = "^.*/srv/.*/requests$".r
-  def mkCounters(metrics: Map[String, Number]): Seq[Counter] =
-    metrics.collect {
-      case (RequestPattern(), v) => Counter(Some("srv_requests"), Some(v.longValue()))
-    }.toSeq
+  def counterValue(metric: Metric): Option[Long] = metric match {
+    case c: Metric.Counter => Some(c.get)
+    case _ => None
+  }
+
+  def mkCounters(metrics: MetricsTree): Seq[Counter] = {
+    val counters = for {
+      router <- metrics.resolve(Seq("rt")).children.values
+      server <- router.resolve(Seq("srv")).children.values
+      counter <- counterValue(server.resolve(Seq("requests")).metric)
+    } yield Counter(Some("srv_requests"), Some(counter))
+    counters.toSeq
+  }
 
   def mkRouters(config: Linker.LinkerConfig): Seq[Router] =
     config.routers.map { r =>
@@ -129,11 +141,11 @@ private[telemeter] object UsageDataTelemeter {
     config: Linker.LinkerConfig,
     pid: String,
     orgId: Option[String],
-    registry: Metrics
+    metrics: MetricsTree
   ) extends Service[Request, Response] {
 
     def apply(request: Request): Future[Response] = {
-      val msg: UsageMessage = mkUsageMessage(config, pid, orgId, registry.sample().asScala.toMap)
+      val msg: UsageMessage = mkUsageMessage(config, pid, orgId, metrics)
       val rsp = Response()
       rsp.contentType = MediaType.Json
       rsp.contentString = mapper.writeValueAsString(msg)
@@ -159,7 +171,7 @@ class UsageDataTelemeter(
   metricsDst: Name,
   withTls: Boolean,
   config: Linker.LinkerConfig,
-  registry: com.twitter.common.metrics.Metrics,
+  metrics: MetricsTree,
   orgId: Option[String],
   dryRun: Boolean
 ) extends Telemeter with Admin.WithHandlers {
@@ -175,10 +187,8 @@ class UsageDataTelemeter(
   val client = Client(metricsService, config, pid, orgId)
 
   val adminHandlers = Seq(
-    Handler("/admin/metrics/usage", new UsageDataHandler(metricsService, config, pid, orgId, registry))
+    Handler("/admin/metrics/usage", new UsageDataHandler(metricsService, config, pid, orgId, metrics))
   )
-
-  private[this] def sample(registry: Metrics): Map[String, Number] = registry.sample().asScala.toMap
 
   // Only run at most once.
   def run(): Closable with Awaitable[Unit] =
@@ -186,10 +196,10 @@ class UsageDataTelemeter(
     else Telemeter.nopRun
 
   private[this] def run0() = {
-    val _ = client(sample(registry))
+    val _ = client(metrics)
 
     val task = DefaultTimer.twitter.schedule(DefaultPeriod) {
-      val _ = client(sample(registry))
+      val _ = client(metrics)
     }
 
     val closer = Closable.sequence(
@@ -218,7 +228,7 @@ case class UsageDataTelemeterConfig(
       Name.bound(Address("stats.buoyant.io", 443)),
       withTls = true,
       config,
-      com.twitter.common.metrics.Metrics.root,
+      params[MetricsTree],
       orgId,
       dryRun.getOrElse(false)
     )